pyproject.toml
[project]
name = "career-mentor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.6.0",
    "litellm>=1.74.2",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
